Fix incorrect column types
Had wrong types for mysql and postgres migration scripts. Caught during testing on CI server

+++ genie-web/src/main/resources/db/migration/postgresql/V4_3_0_0__Containers.sql
@@ -33,9 +33,9 @@ ALTER TABLE jobs
 
 ALTER TABLE commands
   DROP COLUMN check_delay,
-  ADD COLUMN cpu          INT  DEFAULT NULL,
-  ADD COLUMN gpu          INT  DEFAULT NULL,
-  ADD COLUMN disk_mb      INT  DEFAULT NULL,
-  ADD COLUMN network_mbps INT  DEFAULT NULL,
-  ADD COLUMN images       TEXT DEFAULT NULL,
+  ADD COLUMN cpu          INT    DEFAULT NULL,
+  ADD COLUMN gpu          INT    DEFAULT NULL,
+  ADD COLUMN disk_mb      BIGINT DEFAULT NULL,
+  ADD COLUMN network_mbps BIGINT DEFAULT NULL,
+  ADD COLUMN images       TEXT   DEFAULT NULL,
   ALTER COLUMN memory TYPE BIGINT;
